stm32/stm_device_tree.py: Add pin position table.
"Pin" tags in the ST XML are a bit of a misnomer; they are more of a
logical pin than an actual physical one.  Things like "PA", "PB", "VDD",
"VSS", "VREF+" are all examples of Pin names.  This makes it clear that
the mapping from Pin names to physical positions on the chip footprint
is not 1:1 - they are not even 1:* or *:1.

For an example where the same Pin name is mapped to multiple physical
pins, see STM32H742V(G-I)Hx.xml where we have the following mappings:

    <Pin Name="VDD" Position="D2" Type="Power"/>
    <Pin Name="VDD" Position="F5" Type="Power"/>
    <Pin Name="VDD" Position="K1" Type="Power"/>

Likewise, for an example where different Pin names are mapped to the
same physical pin, see STM32G031J(4-6)Mx.xml where we have the following
mappings:

    <Pin Name="PC14-OSC32_IN (PC14)" Position="1" Type="I/O">
    <Pin Name="PB7" Position="1" Type="I/O">
    <Pin Name="PB8" Position="1" Type="I/O">
    <Pin Name="PB9" Position="1" Type="I/O">

The latter case is a small-outline footprint and the SYSCFG component is
used to select which logical Pin is actually active on the physical pin
at any given time.

The good thing about this is that when there are non 1:1 mappings in the
XML, at least we get a separate Pin tag for each mapping instead of
having to do parsing of something like "PB7-PB8-PB9-PC14-OSC32_IN
(PC14)" which would be impractical.

The mapping of logical names to physical pins is useful for tasks such
as board layout tools.  By recording it in a separate table in the
generated XML files, we avoid duplicating thousands of GPIO tags which
would all be identical except for a pin position attribute.