stm32/stm_device_tree.py: For pins which are GPIOs, record their physical position in the pin table.
The pin table maps logical pin names to physical pin locations, and it
also records the pin type ("Power", "I/O", "Reset", etc.).  For the GPIO
pins, it would also be nice to provide a mapping into the separate GPIO
table so that design tools could start with a chip footprint, look at a
physical pin and then figure out what GPIO capabilities are present on
that pin.

This commit adds a copy of the GPIO table's "port" and "pin" attributes
to entries in the pin table that have GPIO functionality so that tools
can use them to make a separate query agains the GPIO table to find the
pin's logical characteristics.