[dfg] Change SAM identifier scheme, fix incorrect identifier parsing
SAM devices with the same letter prefix in the identifier are not
necessarily similar or related. For instance, D51 and E51 are closely
related but D21 and D51 devices are not. Therefore, the naming scheme is
changed to not contain a device family prefix anymore.

The new naming scheme is defined as follows:
    SAM     D21     E    15       A    -    M       U
{platform}{series}{pin}{flash}{variant}-{package}{grade}

The device family is now inferred from the series part of the
identifier, e.g. for D51, E51, E53 and E54 the family is "D5x/E5x".

Furthermore, identifiers with two-letter package codes are now parsed
correctly.